Add STM32L152X RCC PLL and FLASH support.  Now compiles errorfree.

+++ configs/stm32ldiscovery/src/stm32_userleds.c
@@ -106,7 +106,7 @@ void stm32_ledinit(void)
 
 void stm32_setled(int led, bool ledon)
 {
-  uint16_t ledcfg;
+  uint32_t ledcfg;
 
   if (led == BOARD_LED1)
     {
